blocks/deploy_lambda_function/requirements.txt
pip==9.0.1
botocore==1.5.75
boto3==1.4.4
PyYAML==3.12
